Đã có search và database các phòng

+++ QLKDPhongTro.DataLayer/Models/RentedRoom.cs
@@ -6,8 +6,8 @@ namespace QLKDPhongTro.DataLayer.Models
     {
         public int MaPhong { get; set; } = 0;
         public string TenPhong { get; set; } = string.Empty;
-        public double DienTich { get; set; } = 0.0;
-        public double GiaCoBan { get; set; } = 0.0;
+        public decimal DienTich { get; set; } = 0.0m;
+        public decimal GiaCoBan { get; set; } = 0.0m;
         public string TrangThai { get; set; } = string.Empty;
         public string GhiChu { get; set; } = string.Empty;
     }
